+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Unreleased Catalog
 
-* Added scittlets.reagent.codemirror component
+* Added scittlets.reagent.codemirror component (#2)
 
 ## Unreleased CLI